+++ Demo/Sources/Demos/DemosRootViewController.swift
@@ -132,6 +132,17 @@ public final class DemosRootViewController : ListViewController
                 )
                 
                 Item(
+                    DemoItem(text: "Toggle Selection"),
+                    selectionStyle: .toggles(),
+                    onSelect : { _ in
+                        print("Selected")
+                    },
+                    onDeselect: { _ in
+                        print("Deselected")
+                    }
+                )
+                
+                Item(
                     DemoItem(text: "Invoices Payment Schedule"),
                     selectionStyle: .selectable(),
                     onSelect : { _ in
@@ -241,6 +252,14 @@ public final class DemosRootViewController : ListViewController
                         self?.push(PagedViewController())
                     }
                 )
+
+                Item(
+                    DemoItem(text: "Center-Snapping Table Layout"),
+                    selectionStyle: .selectable(),
+                    onSelect : { _ in
+                        self?.push(CenterSnappingTableViewController())
+                    }
+                )
                 
                 Item(
                     DemoItem(text: "Carousel-Style Layouts"),
@@ -271,14 +290,6 @@ public final class DemosRootViewController : ListViewController
                     selectionStyle: .selectable(),
                     onSelect : { _ in
                         self?.push(SpacingCustomizationViewController())
-                    }
-                )
-
-                Item(
-                    DemoItem(text: "Retail Grid Layout"),
-                    selectionStyle: .selectable(),
-                    onSelect : { _ in
-                        self?.push(RetailGridViewController())
                     }
                 )
             } header: {