+++ ListableUI/Sources/Appearance.swift
@@ -19,16 +19,22 @@ public struct Appearance : Equatable
     /// The background color for the list.
     @Color public var backgroundColor : UIColor
     
+    /// The tint color of the refresh control.
+    public var refreshControlColor : UIColor?
+    
     /// If the list should display its scroll indicators.
     public var showsScrollIndicators : Bool
         
     /// Creates a new appearance object with the provided options.
     public init(
         backgroundColor : UIColor = Self.defaultBackgroundColor,
+        refreshControlColor : UIColor? = nil,
         showsScrollIndicators : Bool = true,
         configure : (inout Self) -> () = { _ in }
     ) {
         self._backgroundColor = Color(backgroundColor)
+        
+        self.refreshControlColor = refreshControlColor
         
         self.showsScrollIndicators = showsScrollIndicators