+++ CHANGELOG.md
@@ -10,9 +10,67 @@
 
 ### Changed
 
+- `ListHeaderPosition.fixed` will now only apply if the list's top padding is zero, to avoid the header shifting around during layout if the position is changed.
+
 ### Misc
 
 # Past Releases
+
+# [7.1.2] - 2022-09-12
+
+### Removed
+
+- **Revert from 7.0.0**: _"When a section is inserted or removed, and that section has only one item, and no header or footer, the insertion or removal animation for the section's singular item will be used instead."_ This was causing crashes in `initialLayoutAttributesForAppearingItem` and `finalLayoutAttributesForDisappearingItem` due to index path mismatches.
+
+# [7.1.1] - 2022-09-06
+
+### Fixed
+
+- Fixed an issue where `rowSpacing` was used instead of `itemSpacing` in `FlowListLayout`.
+- Fixed SPM-based builds.
+
+# [7.1.0] - 2022-09-06
+
+### Fixed
+
+- Fixed an issue where `rowSpacing` was used instead of `itemSpacing` in `FlowListLayout`.
+
+### Added
+
+- `Behavior.decelerationRate` - Controls the rate at which scrolling decelerates. The default value, `normal`, maintains the status quo.
+
+- `ListPagingBehavior.firstVisibleItemCentered` - When the user stops scrolling, the final offset of the scroll event will be adjusted so that the first visible item is centered within the visible bounds.
+
+
+# [7.0.0] - 2022-08-16
+
+### Added
+
+- Introduces a `.toggles` selection type, to allow a second tap to deselect an item.
+
+- Swipe actions will now dismiss when touching outside of the actively swiped cell to match iOS behavior.
+
+### Removed
+
+- Removed RetailGridListLayout.
+
+### Changed
+
+- When a section is inserted or removed, and that section has only one item, and no header or footer, the insertion or removal animation for the section's singular item will be used instead.
+
+# [6.0.0] - 2022-07-29
+
+### Fixed
+
+- Fixed an issue where `ListHeaderPosition.fixed` would cause the list header to overlap with the refresh control by falling back to `sticky` behavior if there's a refresh control.
+
+### Added
+
+- `SwipeAction` now allows you to provide an `accessibility{Label,Value,Hint}`, and requires either a `title` or `accessibilityLabel`.
+
+### Changed
+
+- The refresh control color has moved to `Appearance` from `RefreshControl`.
 
 # [5.2.1] - 2022-07-21
 
@@ -724,7 +782,12 @@ listActions.scrolling.scrollToSection(
 Earlier releases were ad-hoc and not tracked. To see all changes, please reference [closed PRs on Github](https://github.com/kyleve/Listable/pulls?q=is%3Apr+is%3Aclosed).
 
 
-[Main]: https://github.com/kyleve/Listable/compare/5.2.1...HEAD
+[Main]: https://github.com/kyleve/Listable/compare/7.1.2...HEAD
+[7.1.2]: https://github.com/kyleve/Listable/compare/7.1.1...7.1.2
+[7.1.1]: https://github.com/kyleve/Listable/compare/7.1.0...7.1.1
+[7.1.0]: https://github.com/kyleve/Listable/compare/7.0.0...7.1.0
+[7.0.0]: https://github.com/kyleve/Listable/compare/6.0.0...7.0.0
+[6.0.0]: https://github.com/kyleve/Listable/compare/5.2.1...6.0.0
 [5.2.1]: https://github.com/kyleve/Listable/compare/5.2.0...5.2.1
 [5.2.0]: https://github.com/kyleve/Listable/compare/5.1.0...5.2.0
 [5.1.0]: https://github.com/kyleve/Listable/compare/5.0.1...5.1.0